Put MdChipRemove and MdChip in one file to avoid circular dependency
