+++ src/lib/chips/index.ts
@@ -8,14 +8,12 @@
 
 import {NgModule} from '@angular/core';
 import {MdChipList} from './chip-list';
-import {MdChip} from './chip';
+import {MdChip, MdChipRemove} from './chip';
 import {MdChipInput} from './chip-input';
-import {MdChipRemove} from './chip-remove';
 
 export * from './chip-list';
 export * from './chip';
 export * from './chip-input';
-export * from './chip-remove';
 
 @NgModule({
   imports: [],
